+++ src/chroma-js.d.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2022 WeatherLayers.com
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+declare module 'chroma-js/src/chroma' {
+  import chroma from 'chroma-js';
+  export = chroma;
+}
+
+declare module 'chroma-js/src/generator/mix' {
+}
+
+declare module 'chroma-js/src/generator/scale' {
+}